refactor(local-community): extract pubsub helpers to local-community/pubsub.ts
Moves _listenToIncomingRequests and _providePubsubTopicRoutingCidsIfNeeded
out of the LocalCommunity class into a focused module. Drops 'private' from
pubsubTopicWithfallback, handleChallengeExchange, and
_lastPubsubTopicRoutingProvideAt so the new free functions can access them
via the community instance.
